fix(agent_session): move QueuedActionDto into the domain
An outbound adapter (connection_gateway_realtime) was importing it from
inbound::axum_router, violating the hexagonal-architecture rule that
outbound adapters depend on the domain, not on each other. The type is
domain-owned (both the queue GET response and the realtime queue
snapshot serialize it byte-identically); axum_router re-exports it so
existing importers are unaffected.
